Deprecate Container in _app
