+++ packages/next/pages/_app.tsx
@@ -1,6 +1,12 @@
 import React, { ErrorInfo } from 'react'
 import PropTypes from 'prop-types'
-import { execOnce, loadGetInitialProps, AppContextType, AppInitialProps, AppPropsType } from 'next-server/dist/lib/utils'
+import {
+  execOnce,
+  loadGetInitialProps,
+  AppContextType,
+  AppInitialProps,
+  AppPropsType,
+} from 'next-server/dist/lib/utils'
 import { Router, makePublicRouterInstance } from '../client/router'
 
 export { AppInitialProps }
@@ -13,12 +19,17 @@ export type AppProps<P = {}> = AppPropsType<Router, P>
  * `App` component is used for initialize of pages. It allows for overwriting and full control of the `page` initialization.
  * This allows for keeping state between navigation, custom error handling, injecting additional data.
  */
-async function appGetInitialProps({ Component, ctx }: AppContext): Promise<AppInitialProps> {
+async function appGetInitialProps({
+  Component,
+  ctx,
+}: AppContext): Promise<AppInitialProps> {
   const pageProps = await loadGetInitialProps(Component, ctx)
   return { pageProps }
 }
 
-export default class App<P = {}, CP = P> extends React.Component<P & AppProps<CP>> {
+export default class App<P = {}, CP = P> extends React.Component<
+  P & AppProps<CP>
+> {
   static childContextTypes = {
     router: PropTypes.object,
   }
@@ -49,37 +60,15 @@ export default class App<P = {}, CP = P> extends React.Component<P & AppProps<CP
   }
 }
 
-export class Container extends React.Component {
-  componentDidMount() {
-    this.scrollToHash()
-  }
-
-  componentDidUpdate() {
-    this.scrollToHash()
-  }
-
-  private scrollToHash() {
-    let { hash } = window.location
-    hash = hash && hash.substring(1)
-    if (!hash) return
-
-    const el = document.getElementById(hash)
-    if (!el) return
-
-    // If we call scrollIntoView() in here without a setTimeout
-    // it won't scroll properly.
-    setTimeout(() => el.scrollIntoView(), 0)
-  }
-
-  render() {
-    return this.props.children
-  }
+// @deprecated noop for now until removal
+export function Container(p: any) {
+  return p.children
 }
 
 const warnUrl = execOnce(() => {
   if (process.env.NODE_ENV !== 'production') {
     console.error(
-      `Warning: the 'url' property is deprecated. https://err.sh/zeit/next.js/url-deprecated`,
+      `Warning: the 'url' property is deprecated. https://err.sh/zeit/next.js/url-deprecated`
     )
   }
 })
